Renaming app to computer_vision
