+++ project/api/computer_vision/s3_utils.py
